refactor: make section reusable
