--- web-components/src/components/organisms/EcologicalCreditCardsSection/index.ts
@@ -1,1 +1,0 @@
-export { EcologicalCreditCardsSection as default } from './EcologicalCreditCardsSection';
