+++ web-components/src/components/organisms/Section/Section.mock.ts
@@ -1,0 +1,7 @@
+import { EcologicalCreditCardMock } from 'src/components/molecules/EcologicalCreditCard/EcologicalCreditCard.mock';
+
+export const sectionCardsMock = [
+  { ...EcologicalCreditCardMock },
+  { ...EcologicalCreditCardMock, title: 'Kulshan Carbon Trust' },
+  { ...EcologicalCreditCardMock, title: 'ERA Brazil' },
+];
