+++ web-components/src/components/organisms/Section/index.ts
@@ -1,0 +1,1 @@
+export { Section as default } from './Section';
